refactor(test/domain/Result): Result로 랭킹, 수익률에 대한 책임 이전에 따른 테스트 위치 변경
